requirements.txt
streamlit
sqlite3
face_recognition
pickle
insightface
opencv-python
numpy
Pillow
onnxruntime
cmake
dlib